Renaming some things and moving everything from the visualizations foldesr to table folders
