+++ src/js/components/dashboard/DashboardContent.jsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import WarningsInfoGraphContainer from 'containers/dashboard/graph/WarningsInfoGraphContainer';
 import DashboardSummaryContainer from 'containers/dashboard/DashboardSummaryContainer';
-import DashboardTableContainer from 'containers/dashboard/visualizations/DashboardTableContainer';
+import DashboardTableContainer from 'containers/dashboard/table/DashboardTableContainer';
 
 const DashboardContent = () => (
     <div>
