+++ src/js/components/dashboard/table/DashboardTable.jsx
@@ -6,7 +6,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import DashboardTableHeader from 'components/dashboard/visualizations/DashboardTableHeader';
+import DashboardTableHeader from 'components/dashboard/table/DashboardTableHeader';
 
 const propTypes = {
     results: PropTypes.array,
@@ -44,7 +44,7 @@ const tableHeaders = [
 export default class DashboardTable extends React.Component {
     render() {
         const tableRows = this.props.results.map((row) => (
-            <tr key={`dashboard-table-row-${row.submission_id}-${row.rule_label}`}>
+            <tr key={`dashboard-table-row-${row.submissionId}-${row.ruleLabel}`}>
                 <td>
                     {row.fileLabel}
                 </td>
@@ -52,14 +52,14 @@ export default class DashboardTable extends React.Component {
                     {row.period}
                 </td>
                 <td>
-                    {row.rule_label}
+                    {row.ruleLabel}
                 </td>
                 <td>
-                    {row.instance_count}
+                    {row.instanceCount}
                 </td>
                 <td>
                     <div className="ellipse-box">
-                        {row.rule_description}
+                        {row.ruleDescription}
                     </div>
                 </td>
             </tr>
