+++ src/js/components/dashboard/table/DashboardTablePagination.jsx
